move API related actions to client module
