+++ lib/sentry/client.ex
@@ -1,10 +1,13 @@
 defmodule Sentry.Client do
+  @type parsed_dsn :: {String.t, String.t, Integer.t}
   @sentry_version 5
+
   quote do
     unquote(@sentry_client "sentry-elixir/#{Mix.Project.config[:version]}")
   end
 
   def request(method, url, headers, body) do
+    body = Poison.encode!(body)
     case :hackney.request(method, url, headers, body, []) do
       {:ok, 200, _headers, client} ->
         case :hackney.body(client) do
@@ -38,5 +41,19 @@ defmodule Sentry.Client do
       {"User-Agent", @sentry_client},
       {"X-Sentry-Auth", authorization_header(public_key, private_key)}
     ]
+  end
+
+  @doc """
+  Parses a Sentry DSN which is simply a URI.
+  """
+  @spec parse_dsn!(String.t) :: parsed_dsn
+  def parse_dsn!(dsn) do
+    # {PROTOCOL}://{PUBLIC_KEY}:{SECRET_KEY}@{HOST}/{PATH}{PROJECT_ID}
+    %URI{userinfo: userinfo, host: host, port: port, path: path, scheme: protocol} = URI.parse(dsn)
+    [public_key, secret_key] = String.split(userinfo, ":", parts: 2)
+    {project_id, _} = String.slice(path, 1..-1)
+                      |> Integer.parse
+    endpoint = "#{protocol}://#{host}:#{port}/api/#{project_id}/store/"
+    {endpoint, public_key, secret_key}
   end
 end
